+++ src/septentrio_gnss_driver/communication/communication_core.cpp
@@ -384,10 +384,19 @@ void io_comm_rx::Comm_IO::configureRx()
 
         rest_interval = rest_sec_or_msec + std::to_string(rx_period_rest);
     }
+
+    // Credentials for login
+    if (!settings_->login_user.empty() &&
+        !settings_->login_password.empty())
+        send("login, " + settings_->login_user + ", " + settings_->login_password + " \x0D"); 
     
     // Turning off all current SBF/NMEA output    
     send("sso, all, none, none, off \x0D");
     send("sno, all, none, none, off \x0D");
+
+    // Activate NTP server
+    if (settings_->use_gnss_time)
+        send("sntp, on \x0D");
 
     // Setting the datum to be used by the Rx (not the NMEA output though, which only
     // provides MSL and undulation (by default with respect to WGS84), but not
@@ -516,6 +525,12 @@ void io_comm_rx::Comm_IO::configureRx()
     }
     
     // Setting up NMEA streams
+    {
+		std::stringstream ss;
+
+		ss << "snti, GP" << "\x0D";
+		send(ss.str());
+	}       
 	if (settings_->publish_gpgga)
 	{
 		std::stringstream ss;
@@ -1373,12 +1388,14 @@ bool io_comm_rx::Comm_IO::initializeSerial(std::string port, uint32_t baudrate,
 
 void io_comm_rx::Comm_IO::setManager(const boost::shared_ptr<Manager>& manager)
 {
+    namespace bp = boost::placeholders;
+    
     node_->log(LogLevel::DEBUG, "Called setManager() method");
     if (manager_)
         return;
     manager_ = manager;
     manager_->setCallback(
-        boost::bind(&CallbackHandlers::readCallback, &handlers_, _1, _2, _3));
+        boost::bind(&CallbackHandlers::readCallback, &handlers_, bp::_1, bp::_2, bp::_3));
     node_->log(LogLevel::DEBUG, "Leaving setManager() method");
 }