+++ CHANGELOG.rst
@@ -2,6 +2,16 @@
 Changelog for package septentrio_gnss_driver
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.2.1 (upcoming)
+------------------
+* New Features
+   * Add login credentials
+   * Activate NTP server if use_gnss_time is set to true
+* Improvements
+   * Add NED option to localization
+* Fixes
+   * IMU orientation for ROS axis convention
+
 1.2.0 (2022-04-27)
 ------------------
 * New Features